+++ src/index.js
@@ -1,3 +1,19 @@
-import game from './js/game.js';
+import Phaser from 'phaser';
+import GameScene from './scenes/GameScene.js';
 
-window.addEventListener('load', () => game());
+const config = {
+  type: Phaser.AUTO,
+  width: 800,
+  height: 600,
+  physics: {
+    default: 'arcade',
+    arcade: {
+      gravity: { y: 200 },
+    },
+  },
+  scene: GameScene,
+};
+
+const start = () => new Phaser.Game(config);
+
+window.addEventListener('load', start);
